backend/api/views.py
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework import status
from .models import Conversation, KnowledgeGraph
# You'll need to import your STT, OpenAI, TTS, and Neo4j services

@api_view(['POST'])
def voice_input(request):
    try:
        # Handle audio file from request
        audio_file = request.FILES.get('audio')
        if not audio_file:
            return Response({'error': 'No audio file provided'}, status=status.HTTP_400_BAD_REQUEST)

        # TODO: Implement STT service
        # transcript = stt_service.transcribe(audio_file)

        # Save conversation
        conversation = Conversation.objects.create(
            user_id=request.data.get('user_id'),
            transcript=transcript
        )

        return Response({
            'conversation_id': conversation.id,
            'transcript': transcript
        })
    except Exception as e:
        return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)

@api_view(['POST'])
def chat_response(request):
    try:
        conversation_id = request.data.get('conversation_id')
        user_message = request.data.get('message')
        
        # TODO: Implement OpenAI GPT service
        # response = openai_service.generate_response(user_message)

        # Update conversation
        conversation = Conversation.objects.get(id=conversation_id)
        conversation.transcript += f"\nUser: {user_message}\nAI: {response}"
        conversation.save()

        return Response({'response': response})
    except Exception as e:
        return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)

@api_view(['POST'])
def voice_output(request):
    try:
        text = request.data.get('text')
        if not text:
            return Response({'error': 'No text provided'}, status=status.HTTP_400_BAD_REQUEST)

        # TODO: Implement TTS service
        # audio_data = tts_service.synthesize(text)

        return Response({
            'audio_url': audio_data  # or audio file URL
        })
    except Exception as e:
        return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)

@api_view(['POST'])
def knowledge_graph(request):
    try:
        conversation_id = request.data.get('conversation_id')
        conversation = Conversation.objects.get(id=conversation_id)

        # TODO: Implement Neo4j service
        # graph_data = neo4j_service.extract_and_update(conversation.transcript)

        # Save knowledge graph
        knowledge_graph = KnowledgeGraph.objects.create(
            conversation_id=conversation_id,
            graph_data=graph_data
        )

        return Response({
            'graph_id': knowledge_graph.id,
            'graph_data': graph_data
        })
    except Exception as e:
        return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)