backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
boto3==1.35.91
botocore==1.35.91
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jmespath==1.0.1
neo4j==5.27.0
openai==1.58.1
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
